% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatedSpectrum.r
\docType{data}
\name{AnnotatedSpectrum}
\alias{AnnotatedSpectrum}
\title{AnnotatedSpectrum Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
AnnotatedSpectrum
}
\description{
AnnotatedSpectrum Class
}
\section{Fields}{

\describe{
\item{\code{collisionEnergy}}{}

\item{\code{empty}}{}

\item{\code{intensities}}{}

\item{\code{masses}}{}

\item{\code{maxIntensity}}{}

\item{\code{msLevel}}{}
}}

\keyword{datasets}
