% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CategoryResultSummary.r
\docType{data}
\name{CategoryResultSummary}
\alias{CategoryResultSummary}
\title{CategoryResultSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
CategoryResultSummary
}
\description{
CategoryResultSummary Class
}
\section{Fields}{

\describe{
\item{\code{categoryName}}{}

\item{\code{categoryProbability}}{}
}}

\keyword{datasets}
