% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollisionEnergy.r
\docType{data}
\name{CollisionEnergy}
\alias{CollisionEnergy}
\title{CollisionEnergy Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
CollisionEnergy
}
\description{
CollisionEnergy Class
}
\section{Fields}{

\describe{
\item{\code{maxEnergy}}{}

\item{\code{minEnergy}}{}
}}

\keyword{datasets}
