% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompoundControllerApi.r
\docType{data}
\name{CompoundControllerApi}
\alias{CompoundControllerApi}
\title{CompoundController operations}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
CompoundControllerApi
}
\description{
swagger.CompoundController
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_compound_ids_using_get getCompoundIds


get_compound_using_get getCompound

}
}

\keyword{datasets}
