% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompoundId.r
\docType{data}
\name{CompoundId}
\alias{CompoundId}
\title{CompoundId Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
CompoundId
}
\description{
CompoundId Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{index}}{}

\item{\code{ionMass}}{}

\item{\code{ionType}}{}

\item{\code{msData}}{}

\item{\code{name}}{}

\item{\code{summary}}{}
}}

\keyword{datasets}
