% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompoundMsData.r
\docType{data}
\name{CompoundMsData}
\alias{CompoundMsData}
\title{CompoundMsData Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
CompoundMsData
}
\description{
CompoundMsData Class
}
\section{Fields}{

\describe{
\item{\code{mergedMs1}}{}

\item{\code{mergedMs2}}{}

\item{\code{ms1Spectra}}{}

\item{\code{ms2Spectra}}{}
}}

\keyword{datasets}
