% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompoundSummary.r
\docType{data}
\name{CompoundSummary}
\alias{CompoundSummary}
\title{CompoundSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
CompoundSummary
}
\description{
CompoundSummary Class
}
\section{Fields}{

\describe{
\item{\code{categoryResultSummary}}{}

\item{\code{formulaResultSummary}}{}

\item{\code{structureResultSummary}}{}
}}

\keyword{datasets}
