% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deviation.r
\docType{data}
\name{Deviation}
\alias{Deviation}
\title{Deviation Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
Deviation
}
\description{
Deviation Class
}
\section{Fields}{

\describe{
\item{\code{absolute}}{}

\item{\code{identifier}}{}

\item{\code{ppm}}{}
}}

\keyword{datasets}
