% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileStore.r
\docType{data}
\name{FileStore}
\alias{FileStore}
\title{FileStore Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
FileStore
}
\description{
FileStore Class
}
\section{Fields}{

\describe{
\item{\code{blockSize}}{}

\item{\code{readOnly}}{}

\item{\code{totalSpace}}{}

\item{\code{unallocatedSpace}}{}

\item{\code{usableSpace}}{}
}}

\keyword{datasets}
