% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FileSystem.r
\docType{data}
\name{FileSystem}
\alias{FileSystem}
\title{FileSystem Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
FileSystem
}
\description{
FileSystem Class
}
\section{Fields}{

\describe{
\item{\code{fileStores}}{}

\item{\code{open}}{}

\item{\code{readOnly}}{}

\item{\code{rootDirectories}}{}

\item{\code{separator}}{}

\item{\code{userPrincipalLookupService}}{}
}}

\keyword{datasets}
