% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormulaId.r
\docType{data}
\name{FormulaId}
\alias{FormulaId}
\title{FormulaId Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
FormulaId
}
\description{
FormulaId Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{ionType}}{}

\item{\code{molecularFormula}}{}

\item{\code{resultScores}}{}
}}

\keyword{datasets}
