% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormulaResultControllerApi.r
\docType{data}
\name{FormulaResultControllerApi}
\alias{FormulaResultControllerApi}
\title{FormulaResultController operations}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
FormulaResultControllerApi
}
\description{
swagger.FormulaResultController
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_fingerprint_using_get getFingerprint


get_formula_ids_using_get getFormulaIds


get_formula_result_scores_using_get getFormulaResultScores


get_formula_result_using_get getFormulaResult


get_fragmentation_tree_using_get getFragmentationTree


get_structure_candidates_using_get getStructureCandidates


get_top_hit_candidate_using_get getTopHitCandidate

}
}

\keyword{datasets}
