% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormulaResultScores.r
\docType{data}
\name{FormulaResultScores}
\alias{FormulaResultScores}
\title{FormulaResultScores Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
FormulaResultScores
}
\description{
FormulaResultScores Class
}
\section{Fields}{

\describe{
\item{\code{canopusResult}}{}

\item{\code{confidenceScore}}{}

\item{\code{siriusScore}}{}

\item{\code{topFingerblastScore}}{}

\item{\code{zodiacScore}}{}
}}

\keyword{datasets}
