% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormulaResultSummary.r
\docType{data}
\name{FormulaResultSummary}
\alias{FormulaResultSummary}
\title{FormulaResultSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
FormulaResultSummary
}
\description{
FormulaResultSummary Class
}
\section{Fields}{

\describe{
\item{\code{adduct}}{}

\item{\code{isotopeScore}}{}

\item{\code{medianMassDeviation}}{}

\item{\code{molecularFormula}}{}

\item{\code{numOfexplainablePeaks}}{}

\item{\code{numOfexplainedPeaks}}{}

\item{\code{siriusScore}}{}

\item{\code{totalExplainedIntensity}}{}

\item{\code{treeScore}}{}

\item{\code{zodiacScore}}{}
}}

\keyword{datasets}
