% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterableFileStore.r
\docType{data}
\name{IterableFileStore}
\alias{IterableFileStore}
\title{IterableFileStore Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
IterableFileStore
}
\description{
IterableFileStore Class
}
\keyword{datasets}
