% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterablePath.r
\docType{data}
\name{IterablePath}
\alias{IterablePath}
\title{IterablePath Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
IterablePath
}
\description{
IterablePath Class
}
\keyword{datasets}
