% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectSpaceControllerApi.r
\docType{data}
\name{ProjectSpaceControllerApi}
\alias{ProjectSpaceControllerApi}
\title{ProjectSpaceController operations}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
ProjectSpaceControllerApi
}
\description{
swagger.ProjectSpaceController
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_project_space_using_get getProjectSpace


get_project_spaces_using_get getProjectSpaces


open_project_space_using_post openProjectSpace


open_project_space_using_put openProjectSpace

}
}

\keyword{datasets}
