% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiriusApi.R
\docType{data}
\name{SiriusApi}
\alias{SiriusApi}
\title{SIRIUS REST API Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
SiriusApi
}
\description{
This class \code{SiriusApi} is a subclass of \code{ApiClient}.
An object of this class represents a client object which handles the client-server communication.
It enables the possibility to start and to shutdown the server with the methods \code{start} and \code{shutdown}.
}
\details{
TODO: Section 'Methods'
}
\keyword{datasets}
