% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StructureResultSummary.r
\docType{data}
\name{StructureResultSummary}
\alias{StructureResultSummary}
\title{StructureResultSummary Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
StructureResultSummary
}
\description{
StructureResultSummary Class
}
\section{Fields}{

\describe{
\item{\code{confidenceScore}}{}

\item{\code{csiScore}}{}

\item{\code{inchiKey}}{}

\item{\code{numOfPubMedIds}}{}

\item{\code{similarity}}{}

\item{\code{smiles}}{}

\item{\code{structureName}}{}

\item{\code{xlogP}}{}
}}

\keyword{datasets}
