% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserPrincipalLookupService.r
\docType{data}
\name{UserPrincipalLookupService}
\alias{UserPrincipalLookupService}
\title{UserPrincipalLookupService Class}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
UserPrincipalLookupService
}
\description{
UserPrincipalLookupService Class
}
\keyword{datasets}
