% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VersionInfoControllerApi.r
\docType{data}
\name{VersionInfoControllerApi}
\alias{VersionInfoControllerApi}
\title{VersionInfoController operations}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
VersionInfoControllerApi
}
\description{
swagger.VersionInfoController
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_version_info_using_get getVersionInfo

}
}

\keyword{datasets}
